wrapped useAuth hook logic into AuthContext
